Comprove test coverage
